requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
cryptography==37.0.4
Django==4.0.6
django-ninja==0.19.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.7.10
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
outcome==1.2.0
pycparser==2.21
pydantic==1.9.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
torch==1.13.0
torchvision==0.14.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
webdriver-manager==3.8.2
wincertstore==0.2
wsproto==1.1.0
